# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend service for JuanitaMusic discord bot",
        "title": "JuanitaAPI",
        "contact": {},
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/search": {
            "get": {
                "tags": [
                    "Search"
                ],
                "summary": "Get all searches",
                "operationId": "GetSearches",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Limit of results",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Search"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Search"
                ],
                "summary": "Create a new search",
                "operationId": "CreateSearch",
                "parameters": [
                    {
                        "description": "Search to create",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Search"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Search"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/search/guild/{guildId}": {
            "get": {
                "tags": [
                    "Search"
                ],
                "summary": "Get all searches by guild",
                "operationId": "GetSearchesByGuild",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Guild ID",
                        "name": "guildId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Limit of results",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Search"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/search/mostplayed": {
            "get": {
                "tags": [
                    "Search"
                ],
                "summary": "Get most played searches",
                "operationId": "GetMostPlayedSearches",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Limit of results",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.SongCount"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/search/mostplayed/guild/{guildId}": {
            "get": {
                "tags": [
                    "Search"
                ],
                "summary": "Get most played searches by guild",
                "operationId": "GetMostPlayedSearchesByGuild",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Guild ID",
                        "name": "guildId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Limit of results",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.SongCount"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/search/mostplayed/requestor/{requestorId}": {
            "get": {
                "tags": [
                    "Search"
                ],
                "summary": "Get most played searches by requestor",
                "operationId": "GetMostPlayedSearchesByRequestor",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Requestor ID",
                        "name": "requestorId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Limit of results",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.SongCount"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/search/requestor/{requestorId}": {
            "get": {
                "tags": [
                    "Search"
                ],
                "summary": "Get all searches by requestor",
                "operationId": "GetSearchesByRequestor",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Requestor ID",
                        "name": "requestorId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Limit of results",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Search"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/search/requestor/{requestorId}/guild/{guildId}": {
            "get": {
                "tags": [
                    "Search"
                ],
                "summary": "Get all searches by requestor and guild",
                "operationId": "GetSearchesByRequestorAndGuild",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Requestor ID",
                        "name": "requestorId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Guild ID",
                        "name": "guildId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Limit of results",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Search"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Requestor": {
            "type": "object",
            "required": [
                "tag"
            ],
            "properties": {
                "id": {
                    "type": "string"
                },
                "tag": {
                    "type": "string"
                }
            }
        },
...
        "models.Search": {
            "type": "object"
        },
        "models.SongCount": {
            "type": "object",
            "required": [
                "duration",
                "play_count",
                "requestor",
                "title",
                "url"
            ],
            "properties": {
                "duration": {
                    "type": "integer"
                },
                "play_count": {
                    "type": "integer"
                },
                "requestor": {
                    "$ref": "#/definitions/models.Requestor"
                },
                "title": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        }
    }
}